Fix final fractional share type mismatches
slippage.py: estimate() and record() qty: int → float
iceberg.py: _ActiveIceberg.total_qty/filled_qty: int → float
telegram_bot.py:
- notify_trade_open/notify_buy qty: int → float
- Format: fractional qty shown as "0.1234" not "0.123456789"
  whole shares shown as "5" not "5.0000"

+++ iceberg.py
@@ -75,10 +75,10 @@ TIMEOUT_SEC:      float = float(os.getenv("ICEBERG_TIMEOUT_SEC",     "20"))
 @dataclass
 class _ActiveIceberg:
     ticker:        str
-    total_qty:     int
+    total_qty:     float
     slices_total:  int
     slices_done:   int   = 0
-    filled_qty:    int   = 0
+    filled_qty:    float = 0.0
     avg_price:     float = 0.0
     started_at:    float = field(default_factory=time.time)
     last_slice_at: float = 0.0
